merge latest official master

+++ src/Screen.h
@@ -26,6 +26,7 @@ extern struct statusData status;
 #include "main.h"
 #include <icons.h>
 #include <string.h>
+#include "../lib/FANETLORA/FanetLora.h"
 
 /*
 #define EINK_BUSY     33
@@ -52,7 +53,7 @@ private:
   void drawMainScreen(void);
   void drawWeatherScreen(void);
   void drawFlightTime(int16_t x, int16_t y, int16_t width, int16_t height,uint32_t tTime);
-  void drawValue(int16_t x, int16_t y, int16_t width, int16_t height,float value,uint8_t decimals);
+  void drawValue(int16_t x, int16_t y, int16_t width, int16_t height,float value,unsigned int decimals);
   void drawCompass(int16_t x, int16_t y, int16_t width, int16_t height,float value);
   void drawBatt(int16_t x, int16_t y, int16_t width, int16_t height,uint8_t value);
   void drawSatCount(int16_t x, int16_t y, int16_t width, int16_t height,uint8_t value);
